requirements.txt
streamlit
Pillow
numpy
tensorflow
opencv-python
keras
pywavelets
joblib
